another small fix

+++ core/test.py
@@ -111,7 +111,7 @@ def test(config, model, counter, test_episodes, device, render, save_video=False
                 stack_obs = torch.from_numpy(np.array(stack_obs)).to(device)
 
             with autocast(dtype=torch.bfloat16):
-                network_output = model.initial_inference(stack_obs.float())
+                network_output = model.initial_inference(stack_obs)
             hidden_state_roots = network_output.hidden_state
             reward_hidden_roots = network_output.reward_hidden
             value_prefix_pool = network_output.value_prefix
